+++ src/extension/panelControllers/contractPanelController.ts
@@ -3,7 +3,7 @@ import * as vscode from "vscode";
 import AutoComplete from "../autoComplete";
 import ContractViewRequest from "../../shared/messages/contractViewRequest";
 import ContractViewState from "../../shared/viewState/contractViewState";
-import Log from "../../shared/log";
+import Log from "../util/log";
 import PanelControllerBase from "./panelControllerBase";
 
 const LOG_PREFIX = "ContractPanelController";
@@ -36,7 +36,7 @@ export default class ContractPanelController extends PanelControllerBase<
   onClose() {}
 
   protected async onRequest(request: ContractViewRequest) {
-    Log.log(LOG_PREFIX, `Request: ${JSON.stringify(request)}`);
+    Log.log(LOG_PREFIX, "Request:", request);
     if (request.copyHash) {
       await vscode.env.clipboard.writeText(this.contractHash);
       vscode.window.showInformationMessage(